requirements.txt
gunicorn
gevent
flask
flask_script
flask-cors
flasgger
peewee
pymysql
itsdangerous
PyJWT
requests
pymongo
mongoengine
